Phase 7: reach zero-failure RC validation
441 checks, 0 failed, agreeing across three builds:

    Release        441 / 0 failed
    Debug          441 / 0 failed
    ASan + UBSan   441 / 0 failed, 0 sanitizer findings

The matrix widened to every rate and block size the brief names - 44.1, 48,
88.2, 96, 176.4 and 192 kHz against blocks of 1, 16, 32, 64, 128, 256, 512,
1024 and 2048, mono and stereo - and every configuration still produces finite
audio at a sane level with no audio-thread allocation.

Frozen invariants all hold unchanged:

    crossover null            -111.9 dB worst sample, -109.0 dB RMS
    latency                   65 samples at every quality
    Drive 0                   THD+N -124.3 dB, gain error 0.0016 dB
    powered-off bands         0.002 dB from Drive 0
    Space switch-on leakage   -66.8 to -71.4 dB across the four engines

auval PASSES on the built AU.

Two -Werror failures in the plug-in targets had to be fixed first: the console
test target does not carry juce_recommended_warning_flags, so -Wfloat-equal and
-Wimplicit-int-float-conversion only fire when the VST3/AU/Standalone targets
are built. Direct float comparisons in BandHeaders now go through
juce::exactlyEqual, and MeterColumn's bar width casts before it subtracts.

+++ Source/Ui/MeterColumn.cpp
@@ -126,7 +126,7 @@ namespace fourcolor::ui
 
         //  Two bars, wide enough apart to read L from R.
         const float gap = 3.0f;
-        const float barW = (barsArea.getWidth() - gap) * 0.5f;
+        const float barW = ((float) barsArea.getWidth() - gap) * 0.5f;
 
         for (int c = 0; c < 2; ++c)
         {
